update: go basic authentication and logging middleware
implementation of the basic authentication in go, connection success to establish with the go server. but it fails in the code part, it is not able to extract the correct authorization data from the request also the request logging is not working. I need to work on the fix but for now, I am commiting the changes.